+++ connecteur/seda-ng/lib/XMLFormattage.class.php
@@ -1,6 +1,0 @@
-<?php
-
-/**
- * @deprecated this file is deprectated in 3.1.0 see lib/XMLFormattage
- *
- */